refactor(website): organize product-feature non-visual files
Apply the visual folder structure to the section's shared files:
components/ (Tiles, TileVisual, TileContent, ScrollEntrance, WindowChrome,
RecordTabHeader), data/ (feature-tiles, visuals, card-spans, record-tabs),
types/ (one export per file), and utils/ (format-counter). ProductFeature.tsx
stays as the section shell; the feature-tiles test moves alongside its data.
