+++ packages/twenty-website/src/sections/product-feature/data/card-spans.ts
@@ -1,0 +1,1 @@
+export const CARD_SPANS = [60, 40, 40, 60, 60, 40];
