+++ packages/twenty-website/src/sections/product-feature/types/feature-bullet-icon.ts
@@ -1,0 +1,9 @@
+export type FeatureBulletIcon =
+  | 'book'
+  | 'check'
+  | 'code'
+  | 'edit'
+  | 'eye'
+  | 'search'
+  | 'tag'
+  | 'users';
