+++ packages/twenty-website/src/sections/product-feature/types/feature-bullet.ts
@@ -1,0 +1,8 @@
+import { type MessageDescriptor } from '@lingui/core';
+
+import { type FeatureBulletIcon } from './feature-bullet-icon';
+
+export type FeatureBullet = {
+  icon: FeatureBulletIcon;
+  text: MessageDescriptor;
+};
