+++ packages/twenty-website/src/sections/product-feature/types/feature-visual-key.ts
@@ -1,0 +1,8 @@
+export type FeatureVisualKey =
+  | 'contacts'
+  | 'dashboard'
+  | 'emails'
+  | 'files'
+  | 'import'
+  | 'pipeline'
+  | 'tasks';
