+++ packages/twenty-website/src/sections/product-feature/types/record-tab-label.ts
@@ -1,0 +1,7 @@
+export type RecordTabLabel =
+  | 'Calendar'
+  | 'Emails'
+  | 'Files'
+  | 'Notes'
+  | 'Tasks'
+  | 'Timeline';
