+++ packages/twenty-website/src/sections/product-feature/utils/format-counter.ts
@@ -1,0 +1,7 @@
+function padCounter(value: number): string {
+  return String(value).padStart(2, '0');
+}
+
+export function formatCounter(position: number, totalCount: number): string {
+  return `${padCounter(position)} / ${padCounter(totalCount)}`;
+}
